+++ src/tg2go/bot/handlers/client/menus/category.py
@@ -1,37 +1,35 @@
 from aiogram.filters.callback_data import CallbackData
 from aiogram.types import InlineKeyboardButton, InlineKeyboardMarkup
 
-from tg2go.bot.handlers.client.menus.common import ClientAction, Menu
+from tg2go.bot.handlers.client.menus.common import (
+    ClientAction,
+    ClientPosition,
+    CreateButton,
+    Menu,
+)
 from tg2go.db.models.common.types import CategoryId, GoodId
+from tg2go.services.client.category import ClientCategoryService
 from tg2go.services.client.good import ClientGoodService
 from tg2go.services.client.order import ClientOrderService
 
 
 class CategoryAction(ClientAction):
+    Card = "🛒 Корзина"
     Back = "⬅️ Назад"
 
 
 class CategoryCallbackData(CallbackData, prefix="client.cat"):
     action: CategoryAction
-    category_id: CategoryId
 
 
 class CategoryGoodCallbackData(CallbackData, prefix="client.cat.good"):
-    category_id: CategoryId
     good_id: GoodId
-
-
-def CreateButton(action: ClientAction, category_id: CategoryId) -> InlineKeyboardButton:
-    return InlineKeyboardButton(
-        text=action.value,
-        callback_data=CategoryCallbackData(
-            action=action, category_id=category_id
-        ).pack(),
-    )
 
 
 async def CategoryMenu(chat_id: int, category_id: CategoryId) -> Menu:
     order_srv = await ClientOrderService.Create(chat_id)
+    cat_srv = ClientCategoryService.Create()
+    category = await cat_srv.GetCategory(category_id)
 
     good_srv = ClientGoodService.Create()
     goods = await good_srv.GetAvailableGoods(category_id)
@@ -43,10 +41,7 @@ async def CategoryMenu(chat_id: int, category_id: CategoryId) -> Menu:
         group.append(
             InlineKeyboardButton(
                 text=good.name,
-                callback_data=CategoryGoodCallbackData(
-                    category_id=category_id,
-                    good_id=good.good_id,
-                ).pack(),
+                callback_data=CategoryGoodCallbackData(good_id=good.good_id).pack(),
             )
         )
 
@@ -57,10 +52,11 @@ async def CategoryMenu(chat_id: int, category_id: CategoryId) -> Menu:
     if group:
         buttons.append(group)
 
-    text = await order_srv.GetOrderInfo()
+    text = await order_srv.GetOrderInfo() + ClientPosition.Category(category)
     buttons = [
+        [CreateButton(cb=CategoryCallbackData, action=CategoryAction.Card)],
         *buttons,
-        [CreateButton(action=CategoryAction.Back, category_id=category_id)],
+        [CreateButton(cb=CategoryCallbackData, action=CategoryAction.Back)],
     ]
     markup = InlineKeyboardMarkup(inline_keyboard=buttons)