+++ src/tg2go/bot/handlers/staff/register.py
@@ -1,6 +1,6 @@
 from aiogram import Dispatcher
 
-from tg2go.bot.handlers.staff.callbacks import category, good, panel as stff, settings
+from tg2go.bot.handlers.staff.callbacks import category, good, panel, settings
 from tg2go.bot.handlers.staff.callbacks.category_action import add as category_add
 from tg2go.bot.handlers.staff.callbacks.good_action import add as good_add
 from tg2go.bot.handlers.staff.commands import staff
@@ -9,7 +9,7 @@ from tg2go.bot.handlers.staff.commands import staff
 def RegisterStaffHandlers(dp: Dispatcher) -> None:
     dp.include_routers(
         staff.router,
-        stff.router,
+        panel.router,
         settings.router,
         category.router,
         category_add.router,