+++ src/tg2go/db/models/order_item.py
@@ -57,4 +57,7 @@ class OrderItem(Base, TimestampMixin):
 
     def GetInfoForClient(self) -> str:
         # TODO: insert image_url somehow
-        return f"{self.good.name}, {self.unit_price_rub}₽ × \n{self.quantity}шт. -- {self.unit_price_rub * self.quantity}₽"
+        text = self.good.GetInfoForClient()
+        text += f"\n\n📋Позиция в заказе:\n{self.unit_price_rub}₽ × \n{self.quantity}шт. -- {self.unit_price_rub * self.quantity}₽"
+
+        return text